adding transformation matrix - work in progress

+++ 04-01-opengl/engine/include/model.hpp
@@ -9,9 +9,9 @@ namespace pt
 struct model
 {
     pt::vertex*     vertices   = nullptr;
-    unsigned short  vert_count = 0;
+    size_t          vert_count = 0;
     unsigned short* indices    = nullptr;
-    unsigned short  ind_count  = 0;
+    size_t          ind_count  = 0;
 
     model() = default;
